requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
numpy==1.26.4
Werkzeug==3.0.1
